Docs fixed.

+++ backend/internal/domains/auction/auction_handler.go
@@ -34,17 +34,17 @@ func NewHandler(service AuctionServiceInterface, sched scheduler.SchedulerInterf
 
 // CreateAuction godoc
 //
-//	@Summary		Create Auction
-//	@Description	Creates a new auction with the provided details
-//	@Tags			auction
-//	@Accept			json
-//	@Produce		json
-//	@Param			body	body		CreateAuctionDTO						true	"Auction details"
-//	@Success		201		{object}	sale_offer.RetrieveDetailedSaleOfferDTO	"Created auction"
-//	@Failure		400		{object}	custom_errors.HTTPError					"Bad request"
-//	@Failure		401		{object}	custom_errors.HTTPError					"Unauthorized"
-//	@Router			/auction [post]
-//	@Security		BearerAuth
+// @Summary		Create Auction
+// @Description	Creates a new auction with the provided details
+// @Tags			auction
+// @Accept			json
+// @Produce		json
+// @Param			body	body		CreateAuctionDTO		true	"Auction details"
+// @Success		201		{object}	RetrieveAuctionDTO		"Created auction"
+// @Failure		400		{object}	custom_errors.HTTPError	"Bad request"
+// @Failure		401		{object}	custom_errors.HTTPError	"Unauthorized"
+// @Router			/auction [post]
+// @Security		BearerAuth
 func (h *Handler) CreateAuction(c *gin.Context) {
 	userID, err := auth.GetUserID(c)
 	if err != nil {
@@ -56,7 +56,7 @@ func (h *Handler) CreateAuction(c *gin.Context) {
 		c.JSON(http.StatusBadRequest, custom_errors.NewHTTPError(err.Error()))
 		return
 	}
-	in.UserID = userID
+	in.UserID = (uint)(userID)
 	dto, err := h.service.Create(&in)
 	if err != nil {
 		c.JSON(http.StatusBadRequest, custom_errors.NewHTTPError(err.Error()))
@@ -107,7 +107,7 @@ func (h *Handler) DeleteAuctionByID(c *gin.Context) {
 		c.JSON(http.StatusBadRequest, custom_errors.NewHTTPError(err.Error()))
 		return
 	}
-	err = h.service.Delete(uint(id), userID)
+	err = h.service.Delete(uint(id), uint(userID))
 	if err != nil {
 		c.JSON(http.StatusBadRequest, custom_errors.NewHTTPError(err.Error()))
 		return
@@ -122,10 +122,10 @@ func (h *Handler) DeleteAuctionByID(c *gin.Context) {
 //	@Tags			auction
 //	@Accept			json
 //	@Produce		json
-//	@Param			body	body		UpdateAuctionDTO						true	"Auction details"
-//	@Success		200		{object}	sale_offer.RetrieveDetailedSaleOfferDTO	"Updated auction"
-//	@Failure		400		{object}	custom_errors.HTTPError					"Bad request"
-//	@Failure		401		{object}	custom_errors.HTTPError					"Unauthorized"
+//	@Param			body	body		UpdateAuctionDTO		true	"Auction details"
+//	@Success		200		{object}	RetrieveAuctionDTO		"Updated auction"
+//	@Failure		400		{object}	custom_errors.HTTPError	"Bad request"
+//	@Failure		401		{object}	custom_errors.HTTPError	"Unauthorized"
 //	@Router			/auction [put]
 //	@Security		BearerAuth
 func (h *Handler) UpdateAuction(c *gin.Context) {
@@ -139,7 +139,7 @@ func (h *Handler) UpdateAuction(c *gin.Context) {
 		c.JSON(http.StatusBadRequest, custom_errors.NewHTTPError(err.Error()))
 		return
 	}
-	dto, err := h.service.Update(&auctionInput, userID)
+	dto, err := h.service.Update(&auctionInput, uint(userID))
 	if err != nil {
 		c.JSON(http.StatusBadRequest, custom_errors.NewHTTPError(err.Error()))
 		return
@@ -171,7 +171,7 @@ func (h *Handler) BuyNow(c *gin.Context) {
 		c.JSON(http.StatusBadRequest, custom_errors.NewHTTPError(err.Error()))
 		return
 	}
-	auction, err := h.service.BuyNow(uint(id), userID)
+	auction, err := h.service.BuyNow(uint(id), uint(userID))
 	if err != nil {
 		c.JSON(http.StatusBadRequest, custom_errors.NewHTTPError(err.Error()))
 		return
